fix i2c write prototypes

+++ src/final/ArduCAM.h
@@ -491,19 +491,19 @@ public:
 	int wrSensorRegs16_16(const struct sensor_reg[]);
 	
 	// Read/write 8 bit value to/from 8 bit register address	
-	byte wrSensorReg8_8(int regID, int regDat);
+	byte wrSensorReg8_8(uint8_t regID, uint8_t regDat);
 	byte rdSensorReg8_8(uint8_t regID, uint8_t* regDat);
 	
 	// Read/write 16 bit value to/from 8 bit register address
-	byte wrSensorReg8_16(int regID, int regDat);
+	byte wrSensorReg8_16(uint8_t regID, uint16_t regDat);
 	byte rdSensorReg8_16(uint8_t regID, uint16_t* regDat);
 	
 	// Read/write 8 bit value to/from 16 bit register address
-	byte wrSensorReg16_8(int regID, int regDat);
+	byte wrSensorReg16_8(uint16_t regID, uint8_t regDat);
 	byte rdSensorReg16_8(uint16_t regID, uint8_t* regDat);
 	
 	// Read/write 16 bit value to/from 16 bit register address
-	byte wrSensorReg16_16(int regID, int regDat);
+	byte wrSensorReg16_16(uint16_t regID, uint16_t regDat);
 	byte rdSensorReg16_16(uint16_t regID, uint16_t* regDat);
 
 	void OV2640_set_JPEG_size(uint8_t size);
